LayoutTests/fast/filesystem/op-get-metadata-expected.txt

* Running: GetMetadata
PASS Succeeded: "/".getMetadata()
PASS Succeeded: "/".getDirectory("/a")
PASS Succeeded: "/a".getMetadata()
PASS this.environment["/a.modificationTime"] is >= this.roundedStartDate
PASS Succeeded: "/".getFile("/b")
PASS Succeeded: "/b".getMetadata()
PASS this.environment["/b.modificationTime"] is >= this.roundedStartDate
PASS Succeeded: "/tmp".remove()
PASS this.expectedErrorCode + "" is "1"
PASS this.environment["/a.modificationTime"] is >= this.environment["/.modificationTime"]
PASS this.environment["/b.modificationTime"] is >= this.environment["/.modificationTime"]
PASS this.environment["/b.modificationTime"] is >= this.environment["/a.modificationTime"]
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
